Set "to" be current date if it's missing
Added a IF check to see if the "to" argument is missing (while "from" is given). If it's missing, then "to" will be set to the current date by default.

I guess this is a quite common demand as we always want to know the situation "until now". This also makes this function more "robust".